improve AUTH in a tx
MULTI should be allowed _before_ running AUTH. Maybe another day.
