+++ cmd_connection.go
@@ -66,19 +66,19 @@ func (m *Miniredis) cmdAuth(c *server.Peer, cmd string, args []string) {
 
 	pw := args[0]
 
-	m.Lock()
-	defer m.Unlock()
-	if m.password == "" {
-		c.WriteError("ERR Client sent AUTH, but no password is set")
-		return
-	}
-	if m.password != pw {
-		c.WriteError("ERR invalid password")
-		return
-	}
+	withTx(m, c, func(c *server.Peer, ctx *connCtx) {
+		if m.password == "" {
+			c.WriteError("ERR Client sent AUTH, but no password is set")
+			return
+		}
+		if m.password != pw {
+			c.WriteError("ERR invalid password")
+			return
+		}
 
-	setAuthenticated(c)
-	c.WriteOK()
+		ctx.authenticated = true
+		c.WriteOK()
+	})
 }
 
 // ECHO
