+++ integration/connection_test.go
@@ -82,3 +82,28 @@ func TestSelect(t *testing.T) {
 		succ("EXEC"),
 	)
 }
+
+func TestAuth(t *testing.T) {
+	testAuthCommands(t,
+		"supersecret",
+		fail("PING"),
+		fail("SET", "foo", "bar"),
+		fail("SET"),
+		fail("SET", "foo", "bar", "baz"),
+		fail("GET", "foo"),
+		fail("AUTH", "nosecret"),
+		succ("AUTH", "supersecret"),
+		succ("SET", "foo", "bar"),
+		succ("GET", "foo"),
+	)
+
+	testCommands(t,
+		fail("AUTH", "foo"),
+	)
+
+	testCommands(t,
+		succ("MULTI"),
+		succ("AUTH", "apassword"),
+		succ("EXEC"),
+	)
+}
